requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-webpack-loader==0.6.0
djangorestframework==3.9.4
idna==2.8
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
